app/requirements.txt
altair==5.3.0
ansible-core==2.15.5
attrs==23.1.0
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kaleido==0.2.1
kiwisolver==1.4.5
kubernetes-validate==1.28.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.1
plotly==5.21.0
plotly-express==0.4.1
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.21
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.2
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
recrete-prognosis @ file:///Users/pierreguilbert/Works/git/github/EPFL-ENAC/sxl-recrete-prognosis
referencing==0.35.1
requests==2.32.3
resolvelib==1.0.1
rich==13.7.1
rpds-py==0.13.0
scipy==1.13.0
six==1.16.0
smmap==5.0.1
statsmodels==0.14.2
streamlit==1.35.0
streamlit_carousel==1.0.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
